// src/pages/models/userinfo.ts
export interface UserInfoBD {
    key?: string;
    uid: string;
    email: string;
    genero: string;
    edad: number;
    altura: number;
    peso: number;
    objetivo: string;
    actividad: string;
    Calorias: number;
    proteinas: number;
    grasas: number;
    carbohidratos: number;
  }